@@ -25,8 +25,8 @@ plugins {
     id "com.github.imflog.kafka-schema-registry-gradle-plugin" version "1.1.0"
 }
 
-sourceCompatibility = "14"
-targetCompatibility = "14"
+sourceCompatibility = "15"
+targetCompatibility = "15"
 
 repositories {
     jcenter()
@@ -79,6 +79,7 @@ dependencies {
     testImplementation "org.apache.kafka:kafka-streams-test-utils:2.7.0"
     testImplementation "org.testcontainers:junit-jupiter:1.15.1"
     testImplementation "org.testcontainers:kafka:1.15.1"
+    testImplementation "org.testcontainers:toxiproxy:1.15.1"
 }
 
 protobuf {
